Fix warning

+++ tests/integration/test_opendaq_device_modules/test_native_streaming_modules.cpp
@@ -949,7 +949,7 @@ TEST_F(NativeStreamingModulesTest, StreamDataLowMaxPacketReadCount)
 
 TEST_F(NativeStreamingModulesTest, ParallelRpcCalls)
 {
-    std::vector<size_t> propertyWriteHistory;
+    std::vector<Int> propertyWriteHistory;
 
     auto createServerInstance = [&propertyWriteHistory]()
     {
